moved the menu to the GeneralizedHeader
